server/inference_server/requirements.txt
redis==4.3.6
requests==2.27.1
numpy>=1.21,<1.26
flask==2.0.3
fastapi==0.109.1
websockets==12.0
uvicorn==0.27.0
python-json-logger==2.0.7 
aiohttp==3.9.3
torch==2.1.2 
torchaudio==2.1.2       
transformers==4.36.2
librosa==0.10.1
sentencepiece>=0.1.95
protobuf>=3.20.0
pydub
python-dotenv==1.0.0
langchain==0.1.9
langchain-openai==0.0.5
langchain-community>=0.0.21,<0.1.0
pinecone-client==3.0.2